+++ src/app.rs
@@ -506,7 +506,7 @@ impl eframe::App for App {
                                 egui::FontId::proportional(theme.fonts.ui_font_size),
                             );
 
-                            ui.menu_button("YAAA byOrlov", |ui| {
+                            ui.menu_button("H2Term byOrlov", |ui| {
                                 apply_menu_style(ui, theme.fonts.ui_font_size);
 
                                 if ui.button("ℹ About").clicked() {